# contracts/doorstep/openapi.yaml
openapi: 3.0.3
info:
  title: Doorstep Home Services API
  version: 1.0.0
  description: >-
    Authoritative REST contract for Doorstep (Urban Company style home
    services): the customer surface (Momentum Android app and postbook-ui
    /doorstep), the professional surface (/pro, the doorstep_pro Android app)
    and the admin-internal family (/internal/admin, admin-service tokens only).


    Served by doorstep-service (port 8122, schema `doorstep`) behind the API
    gateway prefix /v1/doorstep (dormant gate DOORSTEP_PUBLIC_ENABLED, pilot
    allowlist DOORSTEP_PILOT_USER_IDS). Pilot city Hyderabad (`HYD`).


    Conventions (binding for every lane):

    - Every response is the shared envelope: success `{"data": <payload>,
      "meta": {"request_id": "..."}}`, failure `{"error": {"code", "message",
      "details"?}, "meta": {...}}`. Schemas below describe `data` unless they
      are named `*Envelope` or `ErrorEnvelope`.

    - Money is integer paise in fields ending `_paise`. Customer prices are
      GST-inclusive; the split is shown as `taxable_paise` + `tax_paise`.

    - Error codes are stable UPPER_SNAKE strings (see `x-doorstep-error-codes`).
      Clients branch on `error.code`, never on `message`.

    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD` in the city's time
      zone (Asia/Kolkata). Times of day are `HH:MM`.

    - Optional values are emitted as explicit `null`, never omitted, on every
...

...
      gateway refuses /internal/ from the edge.


...


...
servers:
...
security:
  - gatewayIdentity: []
...
paths:
...
  /v1/doorstep/pro/jobs/{id}/extras:
    post:
      summary: Propose an extra from the category rate card (or a catalogue add-on for salon)
      operationId: proProposeExtra
      tags: [ProfessionalJob]
      x-lane: A5
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExtraInput'}
      responses:
        '201':
          description: Proposed (customer approves in the app)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Extra'}
        '422':
          description: DOORSTEP_EXTRA_INVALID, DOORSTEP_EXTRAS_NOT_ALLOWED
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/bookings/{id}/photos/{mediaId}:
    get:
      summary: Read private visit evidence as the booking customer or accepted professional
      operationId: visitPhotoBytes
      tags: [Booking]
      x-lane: A5
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: mediaId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Private sanitized image bytes, no-store and nosniff; never a public media URL
          content:
            image/jpeg:
              schema: {type: string, format: binary}
        '404':
          $ref: '#/components/responses/Error'
...
  /v1/doorstep/pro/jobs/{id}/extras/{extraId}:
    delete:
      summary: Withdraw a proposed extra
      operationId: proWithdrawExtra
      tags: [ProfessionalJob]
      x-lane: A5
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/ExtraIdPath'
      responses:
        '204':
          description: Withdrawn
...
components:
  securitySchemes:
    gatewayIdentity:
      type: apiKey
      in: header
      name: X-Verified-User-Id
      description: Stamped by the gateway from the verified JWT (with X-Internal-Service-Key). Never sent by clients directly.
    adminServiceToken:
      type: apiKey
      in: header
      name: X-Service-Authorization
      description: 'Bearer <ed25519 service token>: iss admin-service, aud doorstep, scope [one doorstep:* permission], act <admin user id>, exp <= 60 s.'
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    ExtraIdPath:
      name: extraId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    Error:
      description: Error envelope (see x-doorstep-error-codes)
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    Meta:
      type: object
      properties:
        request_id: {type: string}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, example: DOORSTEP_ADDON_INVALID}
            message: {type: string}
            details:
              type: object
              additionalProperties: true
        meta: {$ref: '#/components/schemas/Meta'}

    # ---------- enums ----------
...
    Extra:
      type: object
      required: [id, booking_id, kind, rate_card_id, addon_id, name, quantity, unit_price_paise, total_paise, status, evidence_media_id, created_at]
      properties:
        id: {type: string, format: uuid}
        booking_id: {type: string, format: uuid}
        kind: {type: string, enum: [rate_card, addon]}
        rate_card_id: {type: string, format: uuid, nullable: true}
        addon_id: {type: string, format: uuid, nullable: true}
        name: {type: string}
        quantity: {type: integer}
        unit_price_paise: {type: integer, format: int64}
        total_paise: {type: integer, format: int64}
        status: {type: string, enum: [proposed, approved, declined, withdrawn, billed]}
        evidence_media_id: {type: string, nullable: true, description: The photo the professional attached to the extra}
        created_at: {type: string, format: date-time}
    ExtraInput:
      type: object
      required: [quantity]
      properties:
        rate_card_id: {type: string, format: uuid}
        addon_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1}
        evidence_media_id: {type: string, format: uuid}
...